fix issue with US region: attempt 2

+++ backend/provisioning-scripts/download-data.sh
@@ -6,20 +6,22 @@ today_date=$(date +%Y%m%d)
 
 if [ "$5" == "us" -a $1 -lt 21 ]; then
         echo "US REGION"
-        wget https://download.geofabrik.de/$7/$5-midwest-$10101.osm.pbf  -O /tmp/$5-midwest-$10101.osm.pbf
-        wget https://download.geofabrik.de/$7/$5-northeast-$10101.osm.pbf  -O /tmp/$5-northeast-$10101.osm.pbf
-        wget https://download.geofabrik.de/$7/$5-pacific-$10101.osm.pbf  -O /tmp/$5-pacific-$10101.osm.pbf
-        wget https://download.geofabrik.de/$7/$5-south-$10101.osm.pbf  -O /tmp/$5-south-$10101.osm.pbf
-        wget https://download.geofabrik.de/$7/$5-west-$10101.osm.pbf  -O /tmp/$5-west-$10101.osm.pbf
-        
-        osmium merge /tmp/$5-midwest-$10101.osm.pbf /tmp/$5-northeast-$10101.osm.pbf /tmp/$5-pacific-$10101.osm.pbf /tmp/$5-south-$10101.osm.pbf /tmp/$5-west-$10101.osm.pbf -o /tmp/$5-$10101-$4.osm.pbf
-        mv /tmp/$5-$10101-$4.osm.pbf /downloads/$5-$10101.osm.pbf
+        if [ ! -f "/downloads/$5-$10101.osm.pbf" ]; then
+            wget https://download.geofabrik.de/$7/$5-midwest-$10101.osm.pbf  -O /tmp/$5-midwest-$10101.osm.pbf
+            wget https://download.geofabrik.de/$7/$5-northeast-$10101.osm.pbf  -O /tmp/$5-northeast-$10101.osm.pbf
+            wget https://download.geofabrik.de/$7/$5-pacific-$10101.osm.pbf  -O /tmp/$5-pacific-$10101.osm.pbf
+            wget https://download.geofabrik.de/$7/$5-south-$10101.osm.pbf  -O /tmp/$5-south-$10101.osm.pbf
+            wget https://download.geofabrik.de/$7/$5-west-$10101.osm.pbf  -O /tmp/$5-west-$10101.osm.pbf
+            
+            osmium merge /tmp/$5-midwest-$10101.osm.pbf /tmp/$5-northeast-$10101.osm.pbf /tmp/$5-pacific-$10101.osm.pbf /tmp/$5-south-$10101.osm.pbf /tmp/$5-west-$10101.osm.pbf -o /tmp/$5-$10101-$4.osm.pbf
+            mv /tmp/$5-$10101-$4.osm.pbf /downloads/$5-$10101.osm.pbf
 
-        rm /tmp/$5-midwest-$10101.osm.pbf
-        rm /tmp/$5-northeast-$10101.osm.pbf
-        rm /tmp/$5-pacific-$10101.osm.pbf
-        rm /tmp/$5-south-$10101.osm.pbf
-        rm /tmp/$5-west-$10101.osm.pbf
+            rm /tmp/$5-midwest-$10101.osm.pbf
+            rm /tmp/$5-northeast-$10101.osm.pbf
+            rm /tmp/$5-pacific-$10101.osm.pbf
+            rm /tmp/$5-south-$10101.osm.pbf
+            rm /tmp/$5-west-$10101.osm.pbf
+        fi
 
 else
     echo "OTHER REGION"
